Typecasting predictions returned by H5 reader to `float()` within the Callbacks

+++ ahcore/transforms/pre_transforms.py
@@ -175,9 +175,9 @@ def one_hot_encoding(index_map: dict[str, int], mask: np.ndarray) -> np.ndarray:
         One-hot encoded output
     """
     largest_index = max(index_map.values())
-    new_mask = np.zeros((largest_index + 1, *mask.shape))
+    new_mask = np.zeros((largest_index + 1, *mask.shape), dtype=np.float32)
     for idx in range(largest_index + 1):
-        new_mask[idx] = (mask == idx).astype(np.float32)
+        new_mask[idx] = (mask == idx)
     return new_mask
